backend/requirements.txt
Authlib==1.4.0
fastapi==0.115.6
google_api_python_client==2.160.0
google_auth_oauthlib==1.2.1
httpx==0.28.1
Markdown==3.7
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
passlib==1.7.4
Pillow==11.1.0
pydantic==2.10.5
pymongo==3.12.0
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
scikit_learn==1.6.1
slowapi==0.1.9
starlette==0.41.3
uvicorn==0.34.0
